Dev ver

+++ app/src/main/java/com/devidea/timeleft/AppWidget.java
@@ -14,12 +14,12 @@ import androidx.room.Room;
 
 import static com.devidea.timeleft.MainActivity.appDatabase;
 import static com.devidea.timeleft.MainActivity.itemGenerator;
-import static com.devidea.timeleft.MainActivity.timeInfoTime;
-import static com.devidea.timeleft.MainActivity.timeInfoMonth;
-import static com.devidea.timeleft.MainActivity.timeInfoYear;
+import static com.devidea.timeleft.MainActivity.DEFAULT_TIME;
+import static com.devidea.timeleft.MainActivity.DEFAULT_DAY;
+import static com.devidea.timeleft.MainActivity.DEFAULT_YEAR;
 
 public class AppWidget extends AppWidgetProvider {
-    private static final String TAG = "AppWidget";
+
     @Override
     public void onReceive(Context context, Intent intent) {
         super.onReceive(context, intent);
@@ -28,7 +28,7 @@ public class AppWidget extends AppWidgetProvider {
         Log.d("widget", "onReceive() action = " + action);
 
         if (AppWidgetManager.ACTION_APPWIDGET_UPDATE.equals(action)) {
-            if(appDatabase == null){
+            if (appDatabase == null) {
                 appDatabase = Room.databaseBuilder(context, AppDatabase.class, "ItemData")
                         .allowMainThreadQueries()
                         .build();
@@ -80,43 +80,69 @@ public class AppWidget extends AppWidgetProvider {
     @Override
     public void onDeleted(Context context, int[] appWidgetIds) {
         super.onDeleted(context, appWidgetIds);
+        try {
+            appDatabase.DatabaseDao().delete(appWidgetIds[0]);
+        } catch (Exception e) {
+            e.printStackTrace();
+            if (appDatabase == null) {
+                appDatabase = Room.databaseBuilder(context, AppDatabase.class, "ItemData")
+                        .allowMainThreadQueries()
+                        .build();
 
-        appDatabase.DatabaseDao().delete(appWidgetIds[0]);
+                appDatabase.DatabaseDao().delete(appWidgetIds[0]);
+            }
+        }
+
         Log.d("widget", "onDeleted done");
     }
 
     static void updateAppWidget(Context context, AppWidgetManager appWidgetManager, int appWidgetId) {
-        String type = appDatabase.DatabaseDao().getType(appWidgetId);
+        String type = null;
+        try {
+
+            type = appDatabase.DatabaseDao().getType(appWidgetId);
+        } catch (Exception e) {
+            e.printStackTrace();
+            if (appDatabase == null) {
+                appDatabase = Room.databaseBuilder(context, AppDatabase.class, "ItemData")
+                        .allowMainThreadQueries()
+                        .build();
+
+                type = appDatabase.DatabaseDao().getType(appWidgetId);
+            }
+        }
 
         RemoteViews views = new RemoteViews(context.getPackageName(), R.layout.app_widget);
 
+        Intent intentR = new Intent(context, AppWidget.class);
+        intentR.setAction(AppWidgetManager.ACTION_APPWIDGET_UPDATE);
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, intentR, PendingIntent.FLAG_UPDATE_CURRENT);
+        views.setOnClickPendingIntent(R.id.refrash, pendingIntent);
         if (type != null) {
-            Intent intentR = new Intent(context, AppWidget.class);
-            intentR.setAction(AppWidgetManager.ACTION_APPWIDGET_UPDATE);
-            PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, intentR, PendingIntent.FLAG_UPDATE_CURRENT);
-            views.setOnClickPendingIntent(R.id.refrash, pendingIntent);
-
             switch (type) {
                 case "embedYear":
 
-                    views.setTextViewText(R.id.percent_text, timeInfoYear.setTimeItem().getSummery() + timeInfoYear.setTimeItem().getPercentString() + "%");
-                    views.setProgressBar(R.id.progress, 100, (int) Float.parseFloat(timeInfoYear.setTimeItem().getPercentString()), false);
+                    views.setTextViewText(R.id.summery, DEFAULT_YEAR.setTimeItem().getSummery());
+                    views.setTextViewText(R.id.percent, DEFAULT_YEAR.setTimeItem().getPercentString() + "%");
+                    views.setProgressBar(R.id.progress, 100, (int) Float.parseFloat(DEFAULT_YEAR.setTimeItem().getPercentString()), false);
 
                     appWidgetManager.updateAppWidget(appWidgetId, views);
                     break;
 
                 case "embedMonth":
 
-                    views.setTextViewText(R.id.percent_text, timeInfoMonth.setTimeItem().getSummery() + timeInfoMonth.setTimeItem().getPercentString() + "%");
-                    views.setProgressBar(R.id.progress, 100, (int) Float.parseFloat(timeInfoMonth.setTimeItem().getPercentString()), false);
+                    views.setTextViewText(R.id.summery, DEFAULT_DAY.setTimeItem().getSummery());
+                    views.setTextViewText(R.id.percent, DEFAULT_DAY.setTimeItem().getPercentString() + "%");
+                    views.setProgressBar(R.id.progress, 100, (int) Float.parseFloat(DEFAULT_DAY.setTimeItem().getPercentString()), false);
 
                     appWidgetManager.updateAppWidget(appWidgetId, views);
                     break;
 
                 case "embedTime":
 
-                    views.setTextViewText(R.id.percent_text, timeInfoTime.setTimeItem().getSummery() + timeInfoTime.setTimeItem().getPercentString() + "%");
-                    views.setProgressBar(R.id.progress, 100, (int) Float.parseFloat(timeInfoTime.setTimeItem().getPercentString()), false);
+                    views.setTextViewText(R.id.summery, DEFAULT_TIME.setTimeItem().getSummery());
+                    views.setTextViewText(R.id.percent, DEFAULT_TIME.setTimeItem().getPercentString() + "%");
+                    views.setProgressBar(R.id.progress, 100, (int) Float.parseFloat(DEFAULT_TIME.setTimeItem().getPercentString()), false);
 
                     appWidgetManager.updateAppWidget(appWidgetId, views);
                     break;
@@ -131,14 +157,16 @@ public class AppWidget extends AppWidgetProvider {
                     } else {
                         adapterItem = itemGenerator.generateMonthItem(entityItemInfo);
                     }
-                    views.setTextViewText(R.id.percent_text, adapterItem.getSummery() +" "+ adapterItem.getPercentString() + "%");
+                    views.setTextViewText(R.id.summery, adapterItem.getSummery());
+                    views.setTextViewText(R.id.percent, adapterItem.getPercentString() + "%");
                     views.setProgressBar(R.id.progress, 100, (int) Float.parseFloat(adapterItem.getPercentString()), false);
+                    appWidgetManager.updateAppWidget(appWidgetId, views);
 
             }
-            Log.d("widget", type + "update done");
         }
 
-        appWidgetManager.updateAppWidget(appWidgetId, views);
+        Log.d("widget", type + "update done");
+
 
     }
